pyproject.toml
[project]
name = "bit-scope"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "apache-airflow>=2.11.0",
    "asyncpraw>=7.8.1",
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "fredapi>=0.5.2",
    "lxml>=6.0.2",
    "matplotlib>=3.9.4",
    "pandas>=2.3.3",
    "playwright>=1.55.0",
    "pyarrow>=21.0.0",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
]
